Added final version of rest-ejb homework
