+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ear/pom.xml
@@ -2,10 +2,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-    <artifactId>SV_MobileWebshop</artifactId>
-    <groupId>com.mycompany</groupId>
-    <version>1.0-SNAPSHOT</version>
-	</parent>
+        <artifactId>SV_MobileWebshop</artifactId>
+        <groupId>com.mycompany</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
 
     <groupId>com.mycompany</groupId>
     <artifactId>SV_MobileWebshop-ear</artifactId>
@@ -16,7 +16,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		 <sonar.login>vendel.serke</sonar.login>
+        <sonar.login>vendel.serke</sonar.login>
         <sonar.password>servena94</sonar.password>
         <sonar.host.url>http://sonar.codingmentor.xyz</sonar.host.url>
         <sonar.language>java</sonar.language>
