+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/pom.xml
@@ -2,10 +2,10 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-    <artifactId>SV_MobileWebshop</artifactId>
-    <groupId>com.mycompany</groupId>
-    <version>1.0-SNAPSHOT</version>
-	</parent>
+        <artifactId>SV_MobileWebshop</artifactId>
+        <groupId>com.mycompany</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
 
     <groupId>com.mycompany</groupId>
     <artifactId>SV_MobileWebshop-ejb</artifactId>
@@ -17,7 +17,7 @@
     <properties>
         <endorsed.dir>${project.build.directory}/endorsed</endorsed.dir>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<sonar.login>vendel.serke</sonar.login>
+        <sonar.login>vendel.serke</sonar.login>
         <sonar.password>servena94</sonar.password>
         <sonar.host.url>http://sonar.codingmentor.xyz</sonar.host.url>
         <sonar.language>java</sonar.language>
@@ -29,7 +29,6 @@
             <artifactId>junit</artifactId>
             <version>4.12</version>
             <scope>test</scope>
-            <type>jar</type>
         </dependency>
         <dependency>
             <groupId>javax</groupId>
@@ -38,17 +37,11 @@
             <scope>provided</scope>
         </dependency>
         <dependency>
-           <groupId>org.hibernate</groupId>
-           <artifactId>
-               hibernate-validator
-           </artifactId>
-           <version>5.0.0.Final</version>
-       </dependency>
-       <dependency>
-           <groupId>org.hibernate</groupId>
-           <artifactId>hibernate-validator-annotation-processor</artifactId>
-           <version>4.1.0.Final</version>
-       </dependency>
+            <groupId>org.hibernate</groupId>
+            <artifactId>hibernate-validator</artifactId>
+            <version>5.0.0.Final</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
     <build>
