--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/annotation/ValidatorQualifier.java
@@ -1,16 +1,0 @@
-package annotation;
-
-
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import javax.inject.Qualifier;
-
-@Qualifier
-@Target({ElementType.METHOD, ElementType.FIELD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ValidatorQualifier {
-
-}
