--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/constraint/BirthDateConstraint.java
@@ -1,22 +1,0 @@
-package constraint;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-import java.lang.annotation.Target;
-import javax.validation.Constraint;
-import javax.validation.Payload;
-
-
-@Constraint(validatedBy = BirthDateValidator.class)
-@Target({ElementType.TYPE})
-@Retention(RUNTIME)
-public @interface BirthDateConstraint {
-
-    String message() default "{Invalid BirthDate error}";
-
-    Class<?>[] groups() default {};
-
-    Class<? extends Payload>[] payload() default {};
-
-}
