--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/exception/BadRequestException.java
@@ -1,14 +1,0 @@
-package exception;
-
-
-public class BadRequestException extends RuntimeException {
-
-    public BadRequestException(String msg) {
-        super(msg);
-    }
-
-    public BadRequestException() {
-        super();
-    }
-
-}
