--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/exception/BadRequestExceptionMapper.java
@@ -1,17 +1,0 @@
-package exception;
-
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
-
-@Provider
-public class BadRequestExceptionMapper implements ExceptionMapper<BadRequestException> {
-
-    @Override
-    public Response toResponse(BadRequestException exception) {
-        return Response.status(Status.BAD_REQUEST).entity(new ErrorDTO(exception.getMessage())).type(MediaType.APPLICATION_JSON).build();
-    }
-
-}
