--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/exception/ErrorDTO.java
@@ -1,27 +1,0 @@
-package exception;
-
-
-public class ErrorDTO {
-
-    private String errorMessage;
-
-    public ErrorDTO() {
-        //public empty constructor for ErrorDTO
-    }
-
-    public ErrorDTO(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-    
-    
-
-    public String getErrorMessage() {
-        return errorMessage;
-    }
-
-    public void setErrorMessage(String errorMessage) {
-        this.errorMessage = errorMessage;
-    }
-    
-    
-}
