--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/exception/ValidationException.java
@@ -1,13 +1,0 @@
-package exception;
-
-public class ValidationException extends RuntimeException{
-    
-    public ValidationException(String msg) {
-        super(msg);
-    }
-
-    public ValidationException() {
-        super();
-    }
-
-}
