--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/exception/ValidationExceptionMapper.java
@@ -1,18 +1,0 @@
-package exception;
-
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.Response.Status;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
-
-
-@Provider
-public class ValidationExceptionMapper implements ExceptionMapper<ValidationException> {
-
-    @Override
-    public Response toResponse(ValidationException exception) {
-        return Response.status(Status.BAD_REQUEST).entity(new ErrorDTO(exception.getMessage())).type(MediaType.APPLICATION_JSON).build();
-    }
-
-}
