--- homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/interceptor/BeanValidation.java
@@ -1,15 +1,0 @@
-package interceptor;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import javax.interceptor.InterceptorBinding;
-
-
-@InterceptorBinding
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD, ElementType.TYPE})
-public @interface BeanValidation {
-
-}
