+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/sv/mycompany/annotation/ValidatorQualifier.java
@@ -1,0 +1,14 @@
+package sv.mycompany.annotation;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import javax.inject.Qualifier;
+
+@Qualifier
+@Target({ElementType.METHOD, ElementType.FIELD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ValidatorQualifier {
+
+}
