+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/sv/mycompany/exception/BadRequestException.java
@@ -1,0 +1,16 @@
+package sv.mycompany.exception;
+
+import javax.ejb.ApplicationException;
+
+@ApplicationException
+public class BadRequestException extends RuntimeException {
+
+    public BadRequestException(String msg) {
+        super(msg);
+    }
+
+    public BadRequestException() {
+        super();
+    }
+
+}
