+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/sv/mycompany/exception/ErrorDTO.java
@@ -1,0 +1,27 @@
+package sv.mycompany.exception;
+
+
+public class ErrorDTO {
+
+    private String errorMessage;
+
+    public ErrorDTO() {
+        //public empty constructor for ErrorDTO
+    }
+
+    public ErrorDTO(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+    
+    
+
+    public String getErrorMessage() {
+        return errorMessage;
+    }
+
+    public void setErrorMessage(String errorMessage) {
+        this.errorMessage = errorMessage;
+    }
+    
+    
+}
