+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/sv/mycompany/exception/ValidationException.java
@@ -1,0 +1,16 @@
+package sv.mycompany.exception;
+
+import javax.ejb.ApplicationException;
+
+@ApplicationException
+public class ValidationException extends RuntimeException{
+    
+    public ValidationException(String msg) {
+        super(msg);
+    }
+
+    public ValidationException() {
+        super();
+    }
+
+}
