+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/sv/mycompany/producer/LoggingProducer.java
@@ -1,0 +1,14 @@
+package sv.mycompany.producer;
+
+import java.util.logging.Logger;
+import javax.enterprise.inject.Produces;
+import javax.enterprise.inject.spi.InjectionPoint;
+
+
+public class LoggingProducer {
+
+    @Produces
+    public Logger produceLogger(InjectionPoint injectionPoint) {
+        return Logger.getLogger(injectionPoint.getMember().getDeclaringClass().getName());
+    }
+}
