+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/java/sv/mycompany/producer/ValidatorProducer.java
@@ -1,0 +1,16 @@
+package sv.mycompany.producer;
+
+import sv.mycompany.annotation.ValidatorQualifier;
+import javax.enterprise.inject.Produces;
+import javax.validation.Validation;
+import javax.validation.Validator;
+import javax.validation.ValidatorFactory;
+
+public class ValidatorProducer {
+
+    @Produces @ValidatorQualifier
+    public Validator produceValidator() {
+        ValidatorFactory vf = Validation.buildDefaultValidatorFactory();
+        return vf.getValidator();
+    }
+}
