+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-ejb/src/main/resources/ValidationMessages.properties
@@ -1,0 +1,1 @@
+InvalidBirthDate.message=Birth date must be before Registration Date.
