+++ homework-1-rest-ejb/SV_MobileWebshop/SV_MobileWebshop-web/src/main/java/sv/mycompany/exceptionmapper/GeneralExceptionMapper.java
@@ -1,0 +1,24 @@
+package sv.mycompany.exceptionmapper;
+
+import java.util.logging.Level;
+import java.util.logging.Logger;
+import javax.inject.Inject;
+import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
+import javax.ws.rs.ext.ExceptionMapper;
+import javax.ws.rs.ext.Provider;
+import sv.mycompany.exception.ErrorDTO;
+
+
+@Provider
+public class GeneralExceptionMapper implements ExceptionMapper<Throwable> {
+
+    @Inject
+    private  Logger LOGGER;
+
+    @Override
+    public Response toResponse(Throwable throwable) {
+        LOGGER.log(Level.SEVERE, "General Exception", throwable);
+        return Response.status(Response.Status.INTERNAL_SERVER_ERROR).entity(new ErrorDTO(throwable.getMessage() + " - " + throwable.getCause())).type(MediaType.APPLICATION_JSON).build();
+    }
+}
