+++ homework-1-rest-ejb/SV_MobileWebshop/pom.xml
@@ -11,5 +11,4 @@
     <module>SV_MobileWebshop-web</module>
     <module>SV_MobileWebshop-ejb</module>
   </modules>
-  
 </project>
